pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dark-factory"
version = "0.1.0"
description = "Autonomous AI coding pipeline — spec to deployment with no human code review"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "PyGithub>=2.1",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
dark-factory = "factory.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["factory"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.isort]
force-single-line = true

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
